Refactor import statements in Search.tsx, page.tsx, and weatherUtils.ts for consistency and readability
