+++ app/components/Search.tsx
@@ -3,13 +3,13 @@
 import { useState } from "react";
 import { GroupBase, SingleValue } from "react-select";
 import { AsyncPaginate, LoadOptions } from "react-select-async-paginate";
+import { GEO_API_OPTIONS, GEO_API_URL } from "../config";
 import {
   FetchResponseData,
   LoadOptionsResponse,
   SearchData
 } from "../models/apiTypes";
 import { SearchProps } from "../models/componentProps";
-import { GEO_API_OPTIONS, GEO_API_URL } from "../utils/config";
 
 /**
  * Search component for searching cities.
