+++ app/utils/weatherUtils.ts
@@ -1,5 +1,4 @@
 import { format } from "date-fns";
-import { Alert, WeatherItem } from "../models/weatherTypes";
 import {
   DIRECTION_SEGMENTS,
   FAHRENHEIT_BASE,
@@ -17,8 +16,9 @@ import {
   SLICE_END_INDEX,
   THREE_QUARTER_PHASE,
   WIND_DIRECTION_DIVISOR
-} from "./config";
+} from "../config";
 import { CurrentWeatherData } from "../models/componentProps";
+import { Alert, WeatherItem } from "../models/weatherTypes";
 
 /**
  * Returns the moon phase based on the given phase value.
@@ -96,7 +96,7 @@ const getWindDirection = (direction: number): string =>
     "W/NW",
     "NW",
     "N/NW"
-  ][ Math.round(direction / WIND_DIRECTION_DIVISOR) % DIRECTION_SEGMENTS ];
+  ][Math.round(direction / WIND_DIRECTION_DIVISOR) % DIRECTION_SEGMENTS];
 
 /**
  * Converts speed from miles per hour to kilometers per hour.
@@ -222,7 +222,7 @@ const formatDate = (timestamp: number) => {
  */
 const formatWeatherData = (item: WeatherItem, timezone: string) => {
   return [
-    { id: "Forecast", result: item.weather[ 0 ].description },
+    { id: "Forecast", result: item.weather[0].description },
     {
       id: "Dew Point",
       result: `${toCelsius(item.dew_point)}°C | ${item.dew_point}°F`
@@ -235,8 +235,8 @@ const formatWeatherData = (item: WeatherItem, timezone: string) => {
       result:
         item.rain !== undefined && item.rain >= 0
           ? `${item.rain}mm | ${mmToInches(item.rain).toFixed(
-            RAIN_PRECISION
-          )}in`
+              RAIN_PRECISION
+            )}in`
           : "0.00"
     },
     { id: "UV Index", result: item.uvi },
@@ -263,13 +263,13 @@ const formatWeatherData = (item: WeatherItem, timezone: string) => {
 const parseWeatherData = (data: CurrentWeatherData) => ({
   alert: data.alerts,
   clouds: data.current.clouds,
-  dailyHigh: Math.floor(data.daily[ 0 ].temp.max),
-  dailyLow: Math.floor(data.daily[ 0 ].temp.min),
-  description: data.current.weather[ 0 ].description,
+  dailyHigh: Math.floor(data.daily[0].temp.max),
+  dailyLow: Math.floor(data.daily[0].temp.min),
+  description: data.current.weather[0].description,
   dewPoint: Math.floor(data.current.dew_point),
   heatIndex: Math.floor(data.current.feels_like),
   humidity: data.current.humidity,
-  moonPhase: getMoonPhase(data.daily[ 0 ].moon_phase),
+  moonPhase: getMoonPhase(data.daily[0].moon_phase),
   temp: Math.floor(data.current.temp),
   timezone: data.timezone,
   uvi: data.current.uvi,
